Add fake-hwclock
Recently I have experienced that rebooting would cause the Pi to forget what the time was. Even though NTPD was enabled, it would require manually updating the system time via `ntpq -p` and prior to doing this the system date and time would be months out of sync.

Fake HWClock fixes this and it is also used in Raspbian.